backend/requirements.txt
aiohttp>=3.8.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
billiard==4.2.1
celery==5.3.6
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.109.2
h11==0.14.0
idna==3.10
kombu==5.4.2
prompt_toolkit==3.0.50
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
redis==5.0.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.45.2
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.27.1
vine==5.1.0
wcwidth==0.2.13
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
backoff==2.2.1
pydantic-settings>=2.0.0
sentence-transformers==2.5.1
numpy==1.26.4
torch==2.2.1
yt-dlp==2024.3.10
google-genai==0.3.2
python-multipart==0.0.9
bcrypt==4.1.2
youtube-transcript-api==0.6.2
whisper==1.1.10
